requirements.txt
requests>=2.31
pandas>=2.2
pyarrow>=15
python-dotenv>=1.0
tenacity>=8.2
beautifulsoup4>=4.12
lxml>=5.0
openpyxl>=3.1
jellyfish>=1.0
tqdm>=4.66
pdfplumber>=0.11
